requirements.txt
aiohttp==3.11.14
discord.py==2.5.2
matplotlib==3.10.1
numpy==2.2.4
redis==5.2.1
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.12.0
